pyproject.toml
[project]
name = "release2ntfy"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "MelianMiko", email = "support@mmk.pw"},
]
dependencies = ["pydantic>=2.11.7", "python-benedict[yaml]>=0.34.1", "aiohttp>=3.12.15", "tabulate>=0.9.0", "termcolor>=3.1.0", "pyyaml>=6.0.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
